@@ -4,7 +4,7 @@ version: 2.0.0
 homepage: https://github.com/gunschu/jitsi_meet
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.10.0 <2.0.0"
 
 dependencies:
