@@ -21,7 +21,7 @@ class JitsiMeet {
   static final Map<RoomNameConstraintType, RoomNameConstraint>
       defaultRoomNameConstraints = {
     RoomNameConstraintType.MIN_LENGTH: new RoomNameConstraint((value) {
-      return value.trim().length >= 3;
+      return value!.trim().length >= 3;
     }, "Minimum room length is 3"),
 
 //    RoomNameConstraintType.MAX_LENGTH : new RoomNameConstraint(
@@ -30,7 +30,7 @@ class JitsiMeet {
 
     RoomNameConstraintType.ALLOWED_CHARS: new RoomNameConstraint((value) {
       return RegExp(r"^[a-zA-Z0-9-_]+$", caseSensitive: false, multiLine: false)
-          .hasMatch(value);
+          .hasMatch(value!);
     }, "Only alphanumeric, dash, and underscore chars allowed"),
 
 //    RoomNameConstraintType.FORBIDDEN_CHARS : new RoomNameConstraint(
@@ -42,12 +42,11 @@ class JitsiMeet {
   /// A JitsiMeetingListener can be attached to this meeting that will automatically
   /// be removed when the meeting has ended
   static Future<JitsiMeetingResponse> joinMeeting(JitsiMeetingOptions options,
-      {JitsiMeetingListener listener,
-      Map<RoomNameConstraintType, RoomNameConstraint>
+      {JitsiMeetingListener? listener,
+      Map<RoomNameConstraintType, RoomNameConstraint>?
           roomNameConstraints}) async {
-    assert(options != null, "options are null");
     assert(options.room != null, "room is null");
-    assert(options.room.trim().isNotEmpty, "room is empty");
+    assert(options.room!.trim().isNotEmpty, "room is empty");
 
     // If no constraints given, take default ones
     // (To avoid using constraint, just give an empty Map)
@@ -66,7 +65,7 @@ class JitsiMeet {
 
     // Validate serverURL is absolute if it is not null or empty
     if (options.serverURL?.isNotEmpty ?? false) {
-      assert(Uri.parse(options.serverURL).isAbsolute,
+      assert(Uri.parse(options.serverURL!).isAbsolute,
           "URL must be of the format <scheme>://<host>[/path], like https://someHost.com");
     }
 
@@ -75,9 +74,9 @@ class JitsiMeet {
       String serverURL = options.serverURL ?? "https://meet.jit.si";
       String key;
       if (serverURL.endsWith("/")) {
-        key = serverURL + options.room;
+        key = serverURL + options.room!;
       } else {
-        key = serverURL + "/" + options.room;
+        key = serverURL + "/" + options.room!;
       }
 
       _perMeetingListeners.update(key, (oldListener) => listener,
@@ -119,10 +118,10 @@ class JitsiMeet {
       }, onError: (dynamic error) {
         debugPrint('Jitsi Meet broadcast error: $error');
         _listeners.forEach((listener) {
-          if (listener.onError != null) listener.onError(error);
+          if (listener.onError != null) listener.onError!(error);
         });
         _perMeetingListeners.forEach((key, listener) {
-          if (listener.onError != null) listener.onError(error);
+          if (listener.onError != null) listener.onError!(error);
         });
       });
       _hasInitialized = true;
@@ -146,23 +145,23 @@ class JitsiMeet {
       switch (message['event']) {
         case "onConferenceWillJoin":
           if (listener.onConferenceWillJoin != null)
-            listener.onConferenceWillJoin(message: message);
+            listener.onConferenceWillJoin!(message: message);
           break;
         case "onConferenceJoined":
           if (listener.onConferenceJoined != null)
-            listener.onConferenceJoined(message: message);
+            listener.onConferenceJoined!(message: message);
           break;
         case "onConferenceTerminated":
           if (listener.onConferenceTerminated != null)
-            listener.onConferenceTerminated(message: message);
+            listener.onConferenceTerminated!(message: message);
           break;
         case "onPictureInPictureWillEnter":
           if (listener.onPictureInPictureWillEnter != null)
-            listener.onPictureInPictureWillEnter(message: message);
+            listener.onPictureInPictureWillEnter!(message: message);
           break;
         case "onPictureInPictureTerminated":
           if (listener.onPictureInPictureTerminated != null)
-            listener.onPictureInPictureTerminated(message: message);
+            listener.onPictureInPictureTerminated!(message: message);
           break;
       }
     });
@@ -170,32 +169,32 @@ class JitsiMeet {
 
   /// Sends a broadcast to per meeting listeners added during joinMeeting
   static void _broadcastToPerMeetingListeners(message) {
-    String url = message['url'];
-    final listener = _perMeetingListeners[url];
+    String? url = message['url'];
+    final listener = _perMeetingListeners[url!];
     if (listener != null) {
       switch (message['event']) {
         case "onConferenceWillJoin":
           if (listener.onConferenceWillJoin != null)
-            listener.onConferenceWillJoin(message: message);
+            listener.onConferenceWillJoin!(message: message);
           break;
         case "onConferenceJoined":
           if (listener.onConferenceJoined != null)
-            listener.onConferenceJoined(message: message);
+            listener.onConferenceJoined!(message: message);
           break;
         case "onConferenceTerminated":
           if (listener.onConferenceTerminated != null)
-            listener.onConferenceTerminated(message: message);
+            listener.onConferenceTerminated!(message: message);
 
           // Remove the listener from the map of _perMeetingListeners on terminate
           _perMeetingListeners.remove(listener);
           break;
         case "onPictureInPictureWillEnter":
           if (listener.onPictureInPictureWillEnter != null)
-            listener.onPictureInPictureWillEnter(message: message);
+            listener.onPictureInPictureWillEnter!(message: message);
           break;
         case "onPictureInPictureTerminated":
           if (listener.onPictureInPictureTerminated != null)
-            listener.onPictureInPictureTerminated(message: message);
+            listener.onPictureInPictureTerminated!(message: message);
           break;
       }
     }
@@ -213,8 +212,8 @@ class JitsiMeet {
 }
 
 class JitsiMeetingResponse {
-  final bool isSuccess;
-  final String message;
+  final bool? isSuccess;
+  final String? message;
   final dynamic error;
 
   JitsiMeetingResponse({this.isSuccess, this.message, this.error});
@@ -226,22 +225,22 @@ class JitsiMeetingResponse {
 }
 
 class JitsiMeetingOptions {
-  String room;
-  String serverURL;
-  String subject;
-  String token;
-  bool audioMuted;
-  bool audioOnly;
-  bool videoMuted;
-  String userDisplayName;
-  String userEmail;
-  String userAvatarURL;
-  FeatureFlag featureFlag;
+  String? room;
+  String? serverURL;
+  String? subject;
+  String? token;
+  bool? audioMuted;
+  bool? audioOnly;
+  bool? videoMuted;
+  String? userDisplayName;
+  String? userEmail;
+  String? userAvatarURL;
+  FeatureFlag? featureFlag;
 
   /// Get feature flags Map with keys as String instead of Enum
   /// Useful as an argument sent to the Kotlin/Swift code
-  Map<String, dynamic> getFeatureFlags() =>
-      (featureFlag != null) ? featureFlag.allFeatureFlags() : new HashMap();
+  Map<String?, dynamic> getFeatureFlags() =>
+      (featureFlag != null) ? featureFlag!.allFeatureFlags() : new HashMap();
 
   @override
   String toString() {
